pyproject.toml
[tool.poetry]
name = "parta"
version = "0.1.0"
description = ""
authors = ["Semyon Berezovsky"]

[tool.poetry.dependencies]
python = "3.10.8"
Django = "4.1.2"
djangorestframework = "3.14.0"
psycopg2 = "2.9.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
